replace unsigned with unsigned short

+++ src/components.cpp
@@ -14,13 +14,13 @@ inline const std::vector<size_type>& Agent::get_node_path() const
 }
 
 // incomplete
-void Link::update_period_travel_time(unsigned iter_no)
+void Link::update_period_travel_time(unsigned short iter_no)
 {
     for (auto& v: vdfps)
         v.run_bpr();
 }
 
-inline bool DemandPeriod::contain_iter_no(unsigned iter_no) const
+inline bool DemandPeriod::contain_iter_no(unsigned short iter_no) const
 {
     if (!se)
         return false;
@@ -31,7 +31,7 @@ inline bool DemandPeriod::contain_iter_no(unsigned iter_no) const
     return true;
 }
 
-double DemandPeriod::get_cap_reduction_ratio(size_type link_no, unsigned iter_no) const
+double DemandPeriod::get_cap_reduction_ratio(size_type link_no, unsigned short iter_no) const
 {
     if (!contain_iter_no(iter_no))
         return 1;
